requirements.txt
annotated-types==0.5.0
anyio==3.7.1
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.6.2
certifi==2020.6.20
chardet==4.0.0
click==8.1.7
cmake==3.27.5
colorama==0.4.4
configobj==5.0.6
constantly==15.1.0
contourpy==1.1.1
cryptography==3.4.8
cycler==0.11.0
dbus-python==1.2.18
distlib==0.3.7
distro==1.7.0
dnspython==2.4.2
exceptiongroup==1.1.3
fastapi==0.103.1
filelock==3.12.4
Flask==2.3.3
fonttools==4.42.1
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
kiwisolver==1.4.5
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lit==16.0.6
MarkupSafe==2.1.3
matplotlib==3.8.0
more-itertools==8.10.0
mpmath==1.3.0
netifaces==0.11.0
networkx==3.1
numpy==1.26.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.0
opencv-python==4.8.0.76
packaging==23.1
pandas==2.1.0
pexpect==4.8.0
Pillow==10.0.1
pipenv==2023.9.8
platformdirs==3.10.0
psutil==5.9.5
ptyprocess==0.7.0
py-cpuinfo==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pydantic==2.3.0
pydantic_core==2.6.3
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.8.0
pymongo==4.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
python-dateutil==2.8.2
python-dotenv==1.0.0
python-magic==0.4.24
pytz==2022.1
PyYAML==5.4.1
requests==2.25.1
scipy==1.11.2
seaborn==0.12.2
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
sniffio==1.3.0
ssh-import-id==5.11
starlette==0.27.0
sympy==1.12
systemd-python==234
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
triton==2.0.0
Twisted==22.1.0
typing_extensions==4.7.1
tzdata==2023.3
ubuntu-advantage-tools==8001
ultralytics==8.0.180
unattended-upgrades==0.1
urllib3==1.26.5
virtualenv==20.24.5
wadllib==1.3.6
Werkzeug==2.3.7
zipp==1.0.0
zope.interface==5.4.0
